+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Types of changes
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [3.6.0]
+
+- `Added` new push mode `upsert` to insert or update rows based on primary keys
+
 ## [3.5.0]
 
 - `Added` logging opening DB connexion that use new SafeUrl for display connection URLs without user and password